Always publish through log channels
